Upgrade for rgbds 0.3.6
There wasn't really anything to update for this new version of rgbds,
except the fact that now all symbols referenced through `jr` have become
symbols IMPORTed in object files. This means that we can drop our shitty
`grep`-based workaround, and use this instead, and as a bonus, be able
to find unreferenced local labels. What's there not to like?